Stop re-absolutization of out of flow insets for getComputedStyle()
When we `getComputedStyle()` of an inset property of an out-of-flow
element, if the value is 'auto', percent or calc, we currently redo
the absolutization from the ComputedStyle. This is error-prone, has
known failues in WPT, and makes it impossible to work with anchor
positioning (and especially position fallback).

As we have already absolutized these values during layout, this patch
simply caches them on LayoutResult to be passed to getComputedStyle().

Fixed: 1423493
Fixed: 1451020
Change-Id: Ib31f17304179d5e5b91c9d91c3e97386b904d0a2